requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
click==8.1.7
python-dotenv==1.0.0
google-generativeai==0.3.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
